+++ src/main/scala/draco/gendrake/GenDrake.scala
@@ -10,7 +10,7 @@ object GenDrake extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("GenDrake", _namePackage = Seq ("draco", "gendrake")))
   lazy val dracoType: Type[GenDrake] = Type[GenDrake] (typeDefinition)
 
-  lazy val elementTypeNames: Seq[String] = Seq ("Emission", "Emit", "Emitter")
+  lazy val elementTypeNames: Seq[String] = Seq ("Emit", "Emitter")
 
   lazy val domainType: Domain[GenDrake] = Domain[GenDrake] (typeDefinition)
   lazy val generator: TypeDefinition => String = draco.Drake.emit